requirements.txt
beautifulsoup4
cdsapi
iris
joblib
loguru
numpy
python-dateutil
pyyaml
requests
